requirements.txt
streamlit==1.37.1
pandas==2.2.2
requests==2.32.3
Pillow==10.4.0
bz2file==0.98